Add Travis CI config for automatic builds and static analysis
Adding a .travis.yml file with config for automated builds enables
travis-ci to perform some basic continuous integration.  This does not
replace our existing build and test infrastructure, since that tests
many more combinations and platforms.

When added to the repository, this enables automated builds and tests
for all branches, and coverity analysis of the coverity_scan branch.

Signed-off-by: Steffan Karger <steffan@karger.me>